% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondEntPermute.R
\name{CondEntPermute}
\alias{CondEntPermute}
\title{Permutation test for mutual information}
\usage{
CondEntPermute(
  X,
  Y1,
  Y2,
  method = "emp",
  disc.X = "equalfreq",
  disc.Y1 = "equalfreq",
  disc.Y2 = "equalfreq",
  alpha = 0.6,
  n.sim = 2000,
  ...
)
}
\arguments{
\item{X}{<object; input> A numeric vector.}

\item{method}{<character; proccessing> a method for estimation of entropy. (see help of infotheo- or minerva-package)}

\item{disc.X}{<character; proccessing> a method for discretization of X.}

\item{n.sim}{<numeric; proccessing> The number of simulation.}

\item{...}{Parameters to be passed to the function *minerva::mine*.}

\item{Y}{<object; input> A numeric vector.}

\item{disc.Y}{<character; proccessing> a method for discretization of Y.}

\item{mine.alpha}{<numeric; proccessing> alpha parameter for the mine method. (works only when *method* is "mine")}
}
\description{
Permutation test for mutual information
}
