% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_information_dist.R
\name{dist.mutualInfo}
\alias{dist.mutualInfo}
\title{Distance measured by mutual information}
\usage{
dist.mutualInfo(data, var.subjID = subjID)
}
\description{
Distance measured by mutual information
}
