% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_and_extract.R
\name{dist_and_extract}
\alias{dist_and_extract}
\title{Calculate distance and extract subjects}
\usage{
dist_and_extract(data, tree_phylo = NULL, method = "euclidean",
  type_extr = "Type\%in\%c(3,4)", Name_type = df.Name_type,
  param.UniFrac = c(0, 0.5, 1, "UW"))
}
\arguments{
\item{data}{A matrix to calculate distance}

\item{method}{Metric of distacnce.}

\item{type_extr}{Will be passed to ExploratoryDataAnalysis::extract_type function}

\item{Name_type}{Will be passeed to ExploratoryDataAnalysis::extract_type function}

\item{tree}{A tree object needed when method="UniFrac"}
}
\description{
Calculate distance and extract subjects
}
