% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_dist.R
\name{extract_type}
\alias{extract_type}
\title{dist-class object manupulation: Extract subjects}
\usage{
extract_type(dist, Type_extr, Name_type)
}
\arguments{
\item{dist}{<object; input> dist-class object with IDs as rownames and colnames}

\item{Type_extr}{<character strings; proccessing> condition to extract, e.g., "!is.na(var_name)"}

\item{Name_type}{<object; input> A data.frame-class object with 2 column: IDs and the variable used to extract IDs}
}
\description{
dist-class object manupulation: Extract subjects
}
