% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitzig.select.R
\name{fitzig.select}
\alias{fitzig.select}
\title{Conduct fitZIG analysis (metagenomeSeq paclage) with selected subpopulation}
\usage{
fitzig.select(
  obj,
  var.select = Type,
  condition = c(3, 4),
  var.analyze = "ACPA",
  rareTrimming,
  fit.zig.settings
)
}
\arguments{
\item{obj}{<object; input> A metagenomeSeq-class object.}

\item{var.select}{<object> A variable name for selection of subpopulation.}

\item{condition}{Levels in var.select variable of to be selected sub population.}

\item{var.analyze}{<character> A variable name for independent variable in analysis model.}

\item{rareTrimming}{<numeric> Taxa with the lesser number of subjects in which the taxa were detected are trimmed.}

\item{fit.zig.settings}{<numeric> Pass to fitZig function.}
}
\description{
Conduct fitZIG analysis (metagenomeSeq paclage) with selected subpopulation
}
