% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling.R
\name{func.tilemap}
\alias{func.tilemap}
\title{Make heatmap-like tile plot. (wrapper of gplots::heatmap2)}
\usage{
func.tilemap(
  fn.output.pdf = "Tile",
  surfix.maintitle = "Group: ",
  bin = FALSE,
  n.breaks = 20,
  col.bias = 2,
  breaks = NULL,
  row.adjust = 0,
  count_table,
  physi_table,
  filt.Features = NULL,
  unitHierarchy = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species",
    "NoAgg"),
  method.dist.Row = c("manhattan", "euclidean"),
  method.hclust.Row = c("ward.D2"),
  method.dist.Col = c("manhattan", "euclidean"),
  method.hclust.Col = c("ward.D2")
)
}
\arguments{
\item{fn.output.pdf}{<character; output> File name of PDF output.}

\item{surfix.maintitle}{<character; output> Surfix of Main title in each page.}

\item{bin}{<numeric; processing>}

\item{n.breaks}{<numeric; processing; prefix =20>}

\item{col.bias}{<numeric; processing; prefix =2> Colour brewer setting}

\item{breaks}{<numeric>}

\item{row.adjust}{<numeric>}

\item{count_table}{<object; input data; mandatory> A data.frame with features as column and subjects as rows.}

\item{physi_table}{<object; input data> A data.frame with 2 column of a feature and subject ID. This ID must be identical with of count_table.}

\item{filt.Features}{<character strings; input data> Select column of count_table. If NULL, all features are selected.}

\item{unitHierarchy}{<for future update...> c('Kingdom', 'Phylum', 'Class', 'Order',  'Family', 'Genus', 'Species', "NoAgg"),}

\item{method.dist.Row}{= c('manhattan','euclidean'),}

\item{method.hclust.Row}{= c('ward.D2'),}

\item{method.dist.Col}{= c('manhattan','euclidean'),}

\item{method.hclust.Col}{= c('ward.D2')}
}
\description{
Make heatmap-like tile plot. (wrapper of gplots::heatmap2)
}
