% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots_ordering.R
\name{mf.boxplot}
\alias{mf.boxplot}
\title{Make many boxplots from tydy ordering sheet}
\usage{
mf.boxplot(data, ggdata, var.x, var.y, size = 0.5, var.col = NA,
  plot.col = "black", box.col = "gray", str, dn.surfix)
}
\arguments{
\item{data}{<object; input> A data frame with variables (ind, var.x, var.y, trans.y, trans.x, var.col, str, dn.surfix)}

\item{ggdata}{<object; input> A object with ggplot-class}

\item{var.x}{<character; proccessing>}

\item{var.y}{<character; proccessing>}

\item{size}{<numeric; proccessing>}

\item{var.col}{<character; proccessing>}

\item{plot.col}{<character; proccessing>}

\item{box.col}{<character; proccessing>}

\item{str}{<character; proccessing>}

\item{dn.surfix}{<character; output>}
}
\description{
Make many boxplots from tydy ordering sheet
}
