% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.cross_posiSum.R
\name{mf.cross_posiSum}
\alias{mf.cross_posiSum}
\title{Extract rows with rowSum >0, then extract cols with colSum>0.}
\usage{
mf.cross_posiSum(data)
}
\arguments{
\item{data}{A data.frame-class object in which phenotype data were input.}
}
\description{
Extract rows with rowSum >0, then extract cols with colSum>0.
}
