% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.find_select_bias.R
\name{mf.find_select_bias}
\alias{mf.find_select_bias}
\title{Complication matrix.}
\usage{
mf.find_select_bias(obj.rpart, Data, colnames.targ, threshPred = 0.5)
}
\arguments{
\item{data}{A data.frame-class object. Subjects' IDs have to be contained in a column named 'id'.}

\item{col_selector}{A regex string to specify names of cols to be analyzed.}
}
\description{
Complication matrix.
}
