% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.ImpSR.R
\name{mf.LCExtrpbyWideData}
\alias{mf.LCExtrpbyWideData}
\title{Extrapolation the least change}
\usage{
mf.LCExtrpbyWideData(data, col.visit = c("vis.0", "vis.6", "vis.18"),
  ink.col.visit = "vis.6", paper.col.visit = "vis.18", ADS.Saihi,
  ADS.Demo, .Set = c("ITT", "FAS", "SAS", "PPS"),
  val.Alloc = val.Alloc)
}
\arguments{
\item{data}{A data.frame-class object.}

\item{col.visit}{A character vector.}

\item{ink.col.visit}{A character vector.}

\item{ADS.Saihi}{A data.frame-class object.}

\item{ADS.Demo}{A data.frame-class object.}

\item{.Set}{A character vector.}

\item{val.Alloc}{A character vector}

\item{paper.col.vosot}{A character vector.}
}
\description{
Extrapolation the least change
}
