% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.rsamp.R
\name{mf.rsamp.lm}
\alias{mf.rsamp.lm}
\title{Resampling method for significance evaluation of some feature selection method.}
\usage{
mf.rsamp.lm(
  var.x = "BVAS",
  data = pData.obj.aggTaxa.ADS,
  count.table = MRcounts(obj.aggTaxa.ADS),
  ori.count.table = MRcounts(obj.aggTaxa),
  itt.rsamp = itt.rsamp.wilcox,
  func.stat = vegan::diversity,
  func.lm = robustbase::lmrob,
  lm.summary.statistics = "coefficients",
  list.do.call.func.stat = list(index = "simpson"),
  list.do.call.func.lm = list(method = "MM"),
  ...
)
}
\arguments{
\item{var.x}{A character string to be RHS of formula in func.im.}

\item{data}{A data.frame-class object in which phenotype data were input.}

\item{count.table}{A data.frame-class object of selected features.}

\item{ori.count.table}{A data.frame-class object of features without selection.}

\item{itt.rsamp}{The number of itteration.}

\item{func.stat}{A function object to compute summary of features (DEFAULT = vegan::diversity)}

\item{func.lm}{A function object to}

\item{lm.summary.statistics}{= vegan::diversity,}

\item{list.do.call.func.stat}{list of arguments to be input func.stat.}

\item{list.do.call.func.lm}{list of arguments to be input func.stat.}

\item{...}{so far ineffective.}
}
\description{
Resampling method for significance evaluation of some feature selection method.
}
