% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.rsamp.mutualinfo_test.statistic.R
\name{mf.rsamp.mutualinfo_test}
\alias{mf.rsamp.mutualinfo_test}
\title{Resampling test for significance of mutual information between alpha-diversity index and some feature selection method.
This function uses MIPermute internally without permutation (this is controlled by setting *n.sim=1*)}
\usage{
mf.rsamp.mutualinfo_test(var.x = "BVAS", data = pData.obj.aggTaxa.ADS,
  count.table = MRcounts(obj.aggTaxa.ADS),
  ori.count.table = MRcounts(obj.aggTaxa),
  itt.rsamp = itt.rsamp.wilcox, func.stat = vegan::diversity,
  MIPermute.method = "emp", MIPermute.discretize.X = "equalfreq",
  MIPermute.discretize.Y = "equalfreq", MIPermute.alpha = 0.6, ...)
}
\arguments{
\item{var.x}{Column name of phenotype data.}

\item{data}{A data.frame-class object in which phenotype data were input.}

\item{count.table}{A data.frame-class object of selected features.}

\item{ori.count.table}{A data.frame-class object of features without selection.}

\item{itt.rsamp}{= itt.rsamp.wilcox,}

\item{func.stat}{= vegan::diversity,}

\item{...}{Arguments to be passed to a function in func.stat and MIPermute function.}
}
\description{
Resampling test for significance of mutual information between alpha-diversity index and some feature selection method.
This function uses MIPermute internally without permutation (this is controlled by setting *n.sim=1*)
}
