% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.rsamp.wilcox_test.statistic.R
\name{mf.rsamp.wilcox_test.statistic}
\alias{mf.rsamp.wilcox_test.statistic}
\title{Make many boxplots from tidy ordering sheet}
\usage{
mf.rsamp.wilcox_test.statistic(var.x = "Disease",
  data = pData.obj.aggTaxa.ADS,
  count.table = MRcounts(obj.aggTaxa.ADS),
  ori.count.table = MRcounts(obj.aggTaxa),
  itt.rsamp = itt.rsamp.wilcox, func.stat = vegan::diversity, ...)
}
\arguments{
\item{df.ADS}{<object; input> A data frame}

\item{df.input_for_cmp_2levels}{<object; input> A data frame with variables (var.x, var.y, str)}
}
\description{
Make many boxplots from tidy ordering sheet
}
