% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots_ordering.R
\name{mf.scatterplot}
\alias{mf.scatterplot}
\title{Make many scatterplots from tidy ordering sheet}
\usage{
mf.scatterplot(
  data,
  var.x,
  var.y,
  trans.y = c("log10", "identity"),
  trans.x = c("log10", "identity"),
  size = 0.5,
  var.col = NA,
  plot.col = "black",
  line.col = "gray",
  cont.col = FALSE,
  contour.col = "gray",
  contour.alpha = 0.6,
  str,
  dn.surfix,
  betas = NULL,
  ggdata = NULL
)
}
\arguments{
\item{data}{<object; input> A data frame with variables (ind, var.x, var.y, trans.y, trans.x, var.col, str, dn.surfix)}

\item{var.x}{<character; proccessing>}

\item{var.y}{<character; proccessing>}

\item{trans.y}{<character> To be passed to scale_y_continuous(trans= )}

\item{trans.x}{<character> To be passed to scale_x_continuous(trans= )}

\item{size}{<numeric; proccessing>}

\item{var.col}{<character; proccessing>}

\item{plot.col}{<character> "color_color"(to be passed to scale_color_gradient) or "color"}

\item{str}{<character; proccessing>}

\item{dn.surfix}{<character; output>}

\item{betas}{<character; proccessing> A numeric vector for coefficients of regression line}

\item{ggdata}{<object; input> A object with ggplot-class}
}
\description{
Make many scatterplots from tidy ordering sheet
}
