% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.ImpSR.R
\name{mf.SRIntporbyWideData}
\alias{mf.SRIntporbyWideData}
\title{Interpolation.}
\usage{
mf.SRIntporbyWideData(data, col.visit)
}
\arguments{
\item{data}{A data.frame-class object.}

\item{col.visit}{A character vector.}
}
\description{
Interpolation.
}
