% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.survggplot_res.R
\name{mf.survggplot}
\alias{mf.survggplot}
\title{Survival analysis on specified SNPs as arguments}
\usage{
mf.survggplot(
  data = data_01,
  SNPs = c("rs6755005", "rs10428132"),
  gt = list(c("AA", "GA", "GG"), c("GG", "GT", "TT")),
  gt_levels = list(c("AA", "GA", "GG"), c("GG", "GT", "TT")),
  surv = c("time", "status"),
  cov = NULL
)
}
\arguments{
\item{data}{A data.frame-class object with column of SNPs}

\item{SNPs}{A character vector of SNPs to be included as covariates of Cox PH surv model.}

\item{gt}{A list of character vectors to specify genotypes of the SNPs. Each vector for each SNP. The order of the vectors should be co with those of argument 'SNPs'.}

\item{surv}{A charactor vector to specify the colnames of time and status arguments to survival::Surv().}

\item{cov}{A character vector to specify column for covariate.}

\item{ge_levels}{A list of character vectors to specify the levels of the genotypes, e.g., c('AA', 'AG', 'GG') or c(0, 1, 2) or c(0, 0, 1).}

\item{first}{Logical.}
}
\description{
Survival analysis on specified SNPs as arguments
}
