% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot_for_res.fitZig.R
\name{mf.volcano_plot}
\alias{mf.volcano_plot}
\title{Calculate distance and extract subjects}
\usage{
mf.volcano_plot(data, aggTaxonLvl = "Genus", plot.out = FALSE,
  x.margin = 0.25, y.margin = 0.25, ggplot.theme = theme_bw(),
  cbPalette = c("#56B4E9", "#D55E00", "#999999", "#009E73", "#F0E442",
  "#0072B2", "#D55E00", "#CC79A7", "#00AFBB", "#E7B800", "#FC4E07"))
}
\arguments{
\item{data}{A matrix-class object}

\item{aggTaxonLvl}{<character; input>}

\item{plot.out}{FALSE}

\item{x.margin}{default:0.25}

\item{y.margin}{default: 0.25}

\item{ggplot.theme}{default: theme_bw(),}

\item{cbPalette}{}
}
\description{
Calculate distance and extract subjects
}
