% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots_ordering.R
\name{mf.wrap.boxplot}
\alias{mf.wrap.boxplot}
\title{Make many boxplots from tydy ordering sheet}
\usage{
mf.wrap.boxplot(D, data, ggdata, ...)
}
\arguments{
\item{D}{<object; input> A data frame with variables (ind, var.x, var.y, trans.y, trans.x, var.col, str, dn.surfix)}

\item{data}{<object;> A data.frame-class object}

\item{ggdata}{<object; input> A object with ggplot-class}
}
\description{
Make many boxplots from tydy ordering sheet
}
