% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.auto_bivariate.R
\name{MIC_auto_bivariate}
\alias{MIC_auto_bivariate}
\title{Make many boxplots from tidy ordering sheet}
\usage{
MIC_auto_bivariate(df.ADS, df.input, select.level = 1)
}
\arguments{
\item{df.ADS}{<object; input> A data frame}

\item{df.input}{<object; input> A data frame with variables (var.x, var.y, str)}
}
\description{
Make many boxplots from tidy ordering sheet
}
