% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIPermute.R
\name{MIPermute}
\alias{MIPermute}
\title{Permutation test for mutual information}
\usage{
MIPermute(X, Y, method = "emp", disc.X = "equalfreq",
  disc.Y = "equalfreq", n.sim = 2000)
}
\arguments{
\item{X}{<object; input> A numeric vector.}

\item{Y}{<object; input> A numeric vector.}

\item{method}{<character; proccessing> a method for estimation of entropy.}

\item{disc.X}{<character; proccessing> a method for discretization of X.}

\item{disc.Y}{<character; proccessing> a method for discretization of Y.}

\item{n.sim}{<numeric; proccessing> The number of simulation.}
}
\description{
Permutation test for mutual information
}
