% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_plot.R
\name{setting_plot}
\alias{setting_plot}
\title{Preset parameters for plots to use ExploratoryDataAnalysis package.}
\usage{
setting_plot(setID = 1)
}
\arguments{
\item{setID}{<numeric; preset=1>}
}
\description{
Preset parameters for plots to use ExploratoryDataAnalysis package.
}
