% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.tableOne.R
\name{T1_for_factorVar}
\alias{T1_for_factorVar}
\title{Create table of summary statistics for categorical variables.}
\usage{
T1_for_factorVar(df.input, ADS)
}
\arguments{
\item{df.input}{<object; input> A data.frame-class object with column of var.x, var.y, str.}

\item{ADS}{A data.frame-class object of analysis dataset.}
}
\description{
Create table of summary statistics for categorical variables.
}
