% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.tableOne.R
\name{T1_for_numVar}
\alias{T1_for_numVar}
\title{Create table of summary statistics for continuous variables.}
\usage{
T1_for_numVar(df.input, ADS)
}
\arguments{
\item{df.input}{<object; input> A data.frame-class object with column of var.x, var.y, str.}

\item{ADS}{A data.frame-class object of analysis dataset.}
}
\description{
Create table of summary statistics for continuous variables.
}
