% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot_for_res.fitZig.R
\name{volcano_plot_for_res.fitZig}
\alias{volcano_plot_for_res.fitZig}
\title{Calculate distance and extract subjects}
\usage{
volcano_plot_for_res.fitZig(
  res.fitZig.obj.aggTaxon_full = res.fitZig.obj.aggTaxon_full,
  aggTaxonLvl = "Genus"
)
}
\arguments{
\item{res.fitZig.obj.aggTaxon_full}{A matrix to calculate distance}

\item{aggTaxonLvl}{<character; input>}
}
\description{
Calculate distance and extract subjects
}
