% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.auto_binomialier.R
\name{wilcox_test.auto_binomialier}
\alias{wilcox_test.auto_binomialier}
\title{Make many boxplots from tidy ordering sheet}
\usage{
\method{wilcox_test}{auto_binomialier}(df.ADS, df.input_for_cmp_2levels)
}
\arguments{
\item{df.ADS}{<object; input> A data frame}

\item{df.input_for_cmp_2levels}{<object; input> A data frame with variables (var.x, var.y, str)}
}
\description{
Make many boxplots from tidy ordering sheet
}
