% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.LOCF.R
\name{mf.LOCFbyWideData}
\alias{mf.LOCFbyWideData}
\title{Simply do LOCF.}
\usage{
mf.LOCFbyWideData(paper = ADS.measurement, ink = ADS.measurement,
  mold = df.miss.measurement, val.Visit)
}
\arguments{
\item{paper}{<object; input> A data frame}

\item{ink}{<object; input> A data frame}

\item{mold}{<object; input> A data frame}

\item{val.Visit}{<object; proccessing> A numeric vector}
}
\description{
Simply do LOCF.
}
