% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots_ordering.R
\name{mf.scatterplot}
\alias{mf.scatterplot}
\title{Make many scatterplots from tydy ordering sheet}
\usage{
mf.scatterplot(data, ggdata, var.x, var.y, trans.y = c("log10",
  "identity"), trans.x = c("log10", "identity"), var.col, cont.col, str,
  dn.surfix, betas)
}
\arguments{
\item{data}{<object; input> A data frame with variables (ind, var.x, var.y, trans.y, trans.x, var.col, str, dn.surfix)}

\item{ggdata}{<object; input> A object with ggplot-class}

\item{var.x}{<character; proccessing>}

\item{var.y}{<character; proccessing>}

\item{var.col}{<character; proccessing>}

\item{str}{<character; proccessing>}

\item{dn.surfix}{<character; output>}

\item{size}{<numeric; proccessing>}

\item{beta}{<character; proccessing> A numeric vector for coefficients of regression line}
}
\description{
Make many scatterplots from tydy ordering sheet
}
