% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.ImpSR.R
\name{mf.SRIntporbyWideData}
\alias{mf.SRIntporbyWideData}
\title{Interpolation.}
\usage{
mf.SRIntporbyWideData(paper = ADS.measurement, ink = ADS.measurement,
  mold = df.miss.measurement, val.Visit)
}
\arguments{
\item{paper}{<object; input> A data frame.}

\item{ink}{<object; input> A data frame. Typically, same with paper. The columns for visits should be names "vis.INDICATOR-OF-VISITS".}

\item{mold}{<object; input> A data frame. Missing is 1, otherwise 0. The columns for visits should be names "miss.INDICATOR-OF-VISITS".}

\item{val.Visit}{<object; proccessing> A numeric vector. Indicators for Visits.}
}
\description{
Interpolation.
}
