% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICorrBoots.R
\name{MICorrBoots}
\alias{MICorrBoots}
\title{Mutual information with weighted bootstrap}
\usage{
MICorrBoots(X, Y, S = NULL, method = "emp", disc.X = "equalfreq",
  disc.Y = "equalfreq", alpha = 0.6, n.sim = 2000,
  group = ADS$subjID, samp.prob = rep(1/length(unique(ADS$subjID)),
  length(unique(ADS$subjID))), ...)
}
\arguments{
\item{X}{<object; input> A numeric vector.}

\item{Y}{<object; input> A numeric vector.}

\item{method}{<character; proccessing> a method for estimation of entropy. (see help of infotheo- or minerva-package)}

\item{disc.X}{<character; proccessing> a method for discretization of X.}

\item{disc.Y}{<character; proccessing> a method for discretization of Y.}

\item{alpha}{<numeric; proccessing> alpha parameter for the mine method. (works only when *method* is "mine")}

\item{n.sim}{<numeric; proccessing> The number of simulation.}

\item{group}{<vector; proccessing> Ve}

\item{samp.prob}{<numeric vector; proccessing> probability}

\item{...}{Parameters to be passed to the function *minerva::mine*.}
}
\description{
Mutual information with weighted bootstrap
}
\examples{
eg.res.MICorrBoots <- MICorrBoots(
X = c( 1,  1,  2,  3,  5,  2,  2,  3,  1),
Y = c( 1,  1,  2,  3,  5,  1,  1,  1,  2),
S = c('A','A','A','A','A','B','B','B','B'),
method='shrink',
disc.X='none',
disc.Y='none',
alpha = NULL,
n.sim = 50,
group = c('A','A','A','A','A','B','B','B','B'),
samp.prob =
  rep(
   1/length( unique( c('A','A','A','A','A','B','B','B','B'))),
   length( unique( c('A','A','A','A','A','B','B','B','B')))
   )
  )


}
